Moved my page function selector, to a function in a handler folder.
